+++ dsmr_backend/management/commands/development_reset.py
@@ -6,9 +6,11 @@ from django.conf import settings
 from dsmr_backup.models.settings import BackupSettings, DropboxSettings
 from dsmr_notification.models.settings import NotificationSetting
 from dsmr_consumption.models.settings import ConsumptionSettings
+from dsmr_mqtt.models.settings.broker import MQTTBrokerSettings
 from dsmr_mindergas.models.settings import MinderGasSettings
 from dsmr_frontend.models.message import Notification
 from dsmr_api.models import APISettings
+from dsmr_mqtt.models import queue
 
 
 class Command(BaseCommand):
@@ -37,6 +39,10 @@ class Command(BaseCommand):
         NotificationSetting.objects.update(
             notification_service=None, pushover_api_key=None, pushover_user_key=None, prowl_api_key=None
         )
+        MQTTBrokerSettings.objects.update(
+            port=8883, secure=MQTTBrokerSettings.SECURE_CERT_NONE, debug=True, username='user', password='password'
+        )
+        queue.Message.objects.all().delete()
         Notification.objects.update(read=True)
         Notification.objects.create(message='Development reset completed.')