+++ dsmr_frontend/models/message.py
@@ -11,7 +11,7 @@ class Notification(models.Model):
     objects = NotificationManager()
 
     message = models.TextField()
-    redirect_to = models.CharField(max_length=64, null=True, blank=True, default=None)
+    redirect_to = models.CharField(max_length=64, null=True, blank=True, default=None)  # = a Django reverse URL!
     read = models.BooleanField(default=False)
 
     class Meta: