+++ dsmr_mqtt/tests/models/test_settings.py
@@ -1,3 +1,5 @@
+from unittest import mock
+
 from django.test import TestCase
 from django.contrib.admin.sites import site
 
@@ -15,6 +17,15 @@ class TestBrokerSettings(TestCase):
 
     def test_to_string(self):
         self.assertNotEqual(str(self.instance), '{} object'.format(self.instance.__class__.__name__))
+
+    @mock.patch('dsmr_mqtt.apps.AppConfig._on_broker_settings_updated_signal')
+    def test_signal(self, signal_mock):
+        self.assertFalse(broker.MQTTBrokerSettings.objects.filter(restart_required=True).exists())
+
+        self.instance.debug = True
+        self.instance.save()
+
+        self.assertTrue(broker.MQTTBrokerSettings.objects.filter(restart_required=True).exists())
 
 
 class TestRawTelegramSettings(TestCase):