+++ dsmrreader/config/base.py
@@ -201,6 +201,7 @@ DSMRREADER_REST_FRAMEWORK_API_USER = 'api-user'
 # Sleep durations for infinity processes. Update these in your own config if you wish to alter them.
 DSMRREADER_BACKEND_SLEEP = 1
 DSMRREADER_DATALOGGER_SLEEP = 0.5
+DSMRREADER_MQTT_SLEEP = 1
 
 # Whether telegrams are logged, in base64 format. Only required for debugging.
 DSMRREADER_LOG_TELEGRAMS = False